% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsHurst.R
\name{acsHurst}
\alias{acsHurst}
\title{The Hurst or fGn autocorrelation structure}
\usage{
acsHurst(H, lag, var = 1)
}
\arguments{
\item{H}{A scalar indicating the Hurst coefficient.}

\item{lag}{A scalar indicating the maximum lag, up to which the fGn is estimated.}

\item{var}{A scalar indicating the variance of the process. If var!=1 then, the autocovariance structure is returned, instead of the autocorrelation structure.}
}
\value{
A vector of length (lag+1) with the values of fGn autocorrelation structure.
}
\description{
The Hurst or fGn autocorrelation structure
}
\examples{
## Estimate and plot an fGn (i.e., Hurst) structure with H=0.7, up to lag 500.
ACF=acsHurst(H=0.7, lag=500, var=1)
plot(0:(length(ACF)-1), ACF)
}
