% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csCAS.R
\name{csCAS}
\alias{csCAS}
\title{The Cauchy-type autocorrelation structure (CAS)}
\usage{
csCAS(param, lag, var = 1)
}
\arguments{
\item{param}{A two-dimenstional vector, containing the parameters of CAS. First position is for b, the second for k.}

\item{lag}{A scalar indicating the maximum lag, up to which CAS is estimated.}

\item{var}{A scalar indicating the variance of the process. If var!=1 then,the autocovariance structure is returned, instead of the autocorrelation structure.}
}
\value{
A vector of length (lag+1) with the values of CAS autocorrelation structure.
}
\description{
The Cauchy-type autocorrelation structure (CAS)
}
\examples{
##  Estimate and plot a CAS structure, with b=1 and k=0.6, up to lag 500.
ACF=acsCAS(param=c(1, 0.6), lag=500, var=1)
plot(0:(length(ACF)-1), ACF)
}
