% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPeriodic.R
\name{csPeriodic}
\alias{csPeriodic}
\title{A classical periodic autocorrelation structure}
\usage{
csPeriodic(param, lag, var = 1)
}
\arguments{
\item{param}{A two-dimenstional vector, containing the parameters of the periodic autocorrelation structure. First position is period (p), and the second the lengthscale (l).}

\item{lag}{A scalar indicating the maximum lag, up to which the fGn is estimated.}

\item{var}{A scalar indicating the variance of the process. If var!=1 then, the autocovariance structure is returned, instead of the autocorrelation structure.}
}
\value{
A vector of length (lag+1) with the values of the periodic autocorrelation structure.
}
\description{
A classical periodid autocorrelation structure derived by David Mackay.
}
\examples{
## Estimate and plot a periodic autocorrelation structure with parameters, period=12 and
lenghtscale=1, up to lag 50.

ACF=acsPeriodic(param=c(12, 1), lag=50, var=1)
plot(0:(length(ACF)-1), ACF, t='l')
}
