% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disagg_Seq_S.R
\name{Disagg_Seq_S}
\alias{Disagg_Seq_S}
\title{Disaggregation of a coarse level timeseries sequence to a finer level timeseries sequence exhibiting the target marginal distribution and correlation structure (stationary).}
\usage{
Disagg_Seq_S(HLSeries, ARTApar, max.iter, steps, Adjust = T)
}
\arguments{
\item{HLSeries}{A vector, thati is coarse level timeseries sequence specifying the values to disaggreate into a time series sequence of finer level..}

\item{ARTApar}{A list containing the parameters of the model. The list is constructed by the function "EstARTAp".}

\item{max.iter}{A scalar specifying the maximum number of allowed repetitions (parameter of the disaggregation algorithm - typically set between 300-500.).}

\item{steps}{A scalar specifying the number of timesteps of the sequence to generate.}

\item{Adjust}{A logical operator (TRUE or FALSE) specifying whether (TRUE) or not (FALSE) to perfom the proportianal adjusting operation  (parameter of the disaggregation algorithm - typically set to TRUE).}
}
\value{
A list of the 3 generated time series (in vector format):
X: The final time series at the actual domain with the target marginal distribution and correlation structure;
}
\description{
Disaggregation of a coarse level timeseries sequence to a finer level timeseries sequence exhibiting the target marginal distribution and correlation structure (stationary).
}
\examples{
## Disaggregation of a sequence of 30 timesteps of 24-hour Rainfall ammounts (i.e., 1 month) to 10-minute sequence
## The lower level process (i.e., that of 1-minute) is assumed to be a zero-inflated one
## (i.e., using a mixed marginal distribution), with p0=0.95, and a Gamma distribution
## for the continuous part with shape=1 and scale=0.2.
## In this case, the target autocorrelation strucure is from
## the CAS ACS with b=0 and k=0.2.
\dontrun{
FX='qmixed'
PFX=list(p0=0.95, Distr=qgamma, shape=1, scale=2)
maxlag=24
ACFT=acsCAS(param = c(0, 0.2), lag=maxlag)
param=EstARTAp(ACF = ACFT, dist = FX, params = PFX, NatafIntMethod = 'GH')

Sim=SimARTAp(ARTApar = param, burn = 1000, steps = 24*6*30) #24h*60min*100Days
sum(Sim$X[1:(24*6)])
HLSeries=apply(X = matrix(data = Sim$X, ncol=24*6, byrow = 1), MARGIN = 1, FUN = sum); HLSeries[1]

disag=Disagg_Seq_S(HLSeries = HLSeries, ARTApar = param,
                   max.iter = 500, steps = 24*6, Adjust = 1)$X

DisSeries=apply(X = matrix(data = disag, ncol=24*6, byrow = 1), MARGIN = 1, FUN = sum)
}
}
