% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistrStats.R
\name{DistrStats}
\alias{DistrStats}
\title{Estimation of the theoretical Mean, Variance, Skewness and Kurtosis coefficients of a distribution function (from the distribution function).}
\usage{
DistrStats(
  dDistr,
  lb = -Inf,
  ub = Inf,
  subdiv = 90000,
  rel.tol = 10^-8,
  abs.tol = 10^-8,
  ...
)
}
\arguments{
\item{dDistr}{The density function (as a function) of the distribution.}

\item{lb}{A scalar indicating the lower bound of distribution.}

\item{ub}{A scalar indicating the upper bound of distribution.}

\item{subdiv}{A scalar indicating the maximum number of subintervals.}

\item{rel.tol}{A scalar indicating the relative accuracy requested.}

\item{abs.tol}{A scalar indicating the absolute accuracy requested.}

\item{...}{Additional named arguments containing the distribution parameters.}
}
\value{
A 4-dimensional vector containing the theoretical Mean, Variance, Skewness and Kurtosis coefficients of the distribution.
}
\description{
Estimate the theoretical Mean, Variance, Skewness and Kurtosis coefficients of a distribution function, based on the distribution function.
}
\examples{
## Gamma distribution with shape=0.5 and scale=2
DistrStats(dDistr = dgamma, shape=0.5,scale=2,lb = 0)

## Zero-inflated (i.e., mixed) distribution with p0=0.7 and
## continuous part given by Gamma distribution with shape=0.5 and scale=2

DistrStats(dDistr = function(x) dmixed(x,Distr = dgamma, p0=0.7, shape=0.5, scale=2),lb = 0)

}
