% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstARTAp.R
\name{EstARTAp}
\alias{EstARTAp}
\title{Estimation of the auxiliary AR(p) model parameters}
\usage{
EstARTAp(ACF, maxlag = 0, dist, params, NatafIntMethod = "GH",
  NoEval = 9, polydeg = 8, ...)
}
\arguments{
\item{ACF}{A vector with the target autocorrelation structure (including lag-0, i.e., 1).}

\item{maxlag}{A scalar incating the order of the AR(p) model. If maxlag=0, then the order of the model is p=(length(ACF)-1)}

\item{dist}{A string indicating the quantile function of the target marginal distribution (i.e., the ICDF).}

\item{params}{A named list with the parameters of the target distribution.}

\item{NatafIntMethod}{A string ("GH", "Int", or "MC"), indicating the intergation method, to resolve the Nataf integral.}

\item{NoEval}{A scalar indicating (default: 9) the number of evaluation points for the integration methods.}

\item{polydeg}{A scalar indicating the order of the fitted polynomial. If polydeg=0, then another curve is fitted.}

\item{...}{Additional named arguments for the selected "NatafIntMethod" method.}
}
\value{
A list with the parameters of the auxiliary Gaussian AR(p) model.
}
\description{
Estimation of parameters of the auxiliary AR(p) model to simulate the auxiliary Gaussian process.
}
\note{
Avoid the use of the "GH" method (i.e., NatafIntMethod='GH'), when the marginal(s) are discrete.
}
\examples{
## Parameter estimation for a process with zero-inflated (i.e., mixed) marginal distribution,
## with p0=0.9, and a Gamma distribution
## for the continuous part with shape=0.1 and scale=1.
## In this case, the Autocorrelation strucure is a simple AR(1) with rho=0.8.
\dontrun{
ACF=as.vector(ARMAacf(ar = 0.8, lag.max = 100))
fx='qmixed'
pfx=list(Distr=qgamma, p0=0.9, shape=0.1, scale=1)

ARTApar=EstARTAp(ACF=ACF, maxlag=0, dist=fx, params=pfx,
NatafIntMethod ='GH', NoEval=9, polydeg=0)
}
}
