% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstnAR1.R
\name{EstnAR1}
\alias{EstnAR1}
\title{Estimation of the auxiliary n-AR(1) model parameters}
\usage{
EstnAR1(ACF, Ar1Num, dist = "qgamma", params, NatafIntMethod = "GH",
  NoEval = 9, polydeg = 8, ...)
}
\arguments{
\item{ACF}{A vector with the target autocorrelation structure (including lag-0, i.e., 1).}

\item{Ar1Num}{A scalar (>=2) indicating the number (n) of AR(1) models.}

\item{dist}{A string indicating the quantile function of the target marginal distribution (i.e., the ICDF).}

\item{params}{A named list with the parameters of the target distribution.}

\item{NatafIntMethod}{A string ("GH", "Int", or "MC"), indicating the intergation method, to resolve the Nataf integral.}

\item{NoEval}{A scalar indicating (default: 9) the number of evaluation points for the integration methods.}

\item{polydeg}{A scalar indicating the order of the fitted polynomial. If polydeg=0, then another curve is fitted.}

\item{...}{Additional named arguments for the selected "NatafIntMethod" method.}
}
\value{
A list with the parameters of the auxiliary Gaussian n-AR(1) model.
}
\description{
Estimation of parameters of the the sum of n univariate AR(1) models to simulate the auxiliary Gaussian process.
}
\note{
Avoid the use of the "GH" method (i.e., NatafIntMethod='GH'), when the marginal(s) are discrete.
}
\examples{
## Parameter estimation for a process with zero-inflated (i.e., mixed) marginal distribution,
## with p0=0.8, and a Gamma distribution
## for the continuous part with shape=0.5 and scale=1.
## In this case, the Autocorrelation strucure is a CAS ACS with b=2 and k=0.5.
\dontrun{
ACF=acsCAS(param = c(2, 0.5), lag = 500, var = 1)
dist='qmixed'
params=list(Distr=qgamma, p0=0.8, shape=0.5, scale=1)
nAR1param=EstnAR1(ACF = ACF, Ar1Num = 3, dist = dist, params = params,
                  NatafIntMethod = 'GH', NoEval = 9, polydeg = 8)
}
}
