% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstSMARTA.R
\name{EstSMARTA}
\alias{EstSMARTA}
\title{Estimation of the auxiliary SMA model parameters}
\usage{
EstSMARTA(dist, params, ACFs, Cmat, DecoMethod = "cor.smooth",
  FFTLag = 512, NatafIntMethod = "GH", NoEval = 9, polydeg = 8,
  ...)
}
\arguments{
\item{dist}{A k-dimensional string vector indicating the quantile function of the target marginal distribution (i.e., the ICDF).}

\item{params}{A k-dimensional named list with the parameters of the target distributions.}

\item{ACFs}{A k-dimensional list with the target autocorrelation structure (including lag-0, i.e., 1).}

\item{Cmat}{A matrix (k x k) containing the lag-0 cross-correlation coefficients of the processes.}

\item{DecoMethod}{A string indicating the decomposition method, in case of a non-positive definite matrix (options: 'cor.smooth' and 'nearPD')}

\item{FFTLag}{A scalar indicating the length of the Fast Fourrier Transform (required to estimate the internal parameters of SMA model). Default value=512.}

\item{NatafIntMethod}{A string ("GH", "Int", or "MC"), indicating the intergation method, to resolve the Nataf integral.}

\item{NoEval}{A scalar (power of 2) indicating (default: 9) the number of evaluation points for the integration methods.}

\item{polydeg}{A scalar indicating the order of the fitted polynomial. If polydeg=0, then another curve is fitted.}

\item{...}{Additional named arguments for the selected "NatafIntMethod" method.}
}
\value{
A list with the parameters of the auxiliary Gaussian SMA model.
}
\description{
Estimation of parameters of the auxiliary SMA model to simulate the auxiliary Gaussian process.
}
\note{
Avoid the use of the "GH" method (i.e., NatafIntMethod='GH'), when the marginal(s) are discrete.
}
\examples{
## Simulation example of a bivariate process with
## zero-inflated marginal distributions.
# Define the simulation parameters ----------------------------------------
\dontrun{
LAG=2^6
FFTLag=2^7
SMALAG=2^6
steps=2^14

PFXs=list()
FXs=c('qmixed','qmixed')
# Gamma distribution: Gamma(shape=2, rate=1)
PFXs[[1]]=list(Distr=qgamma, p0=0.9, shape=1, scale=1)
# Weibull distribution: Weibull(shape=1, scale=2)
PFXs[[2]]=list(Distr=qweibull, p0=0.85, shape=1, scale=2)

ACFs=list()
ACFs[[1]]=acsCAS(param = c(0.1, 0.6), lag = LAG)
ACFs[[2]]=acsCAS(param = c(0.2, 0.3), lag = LAG)

Cmat=matrix(c(1,0.6,0.6,1), ncol=2, nrow=2)

# Calculate SMARTA’s parameters -------------------------------------------
SMAparam=EstSMARTA(dist = FXs, params = PFXs, ACFs = ACFs,
Cmat = Cmat, DecoMethod = 'cor.smooth',
FFTLag = FFTLag, NatafIntMethod = 'GH', NoEval = 9, polydeg = 8)
}
}
