% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstSPARTA.R
\name{EstSPARTA}
\alias{EstSPARTA}
\title{Estimation of the auxiliary PAR(1) model parameters}
\usage{
EstSPARTA(s2srtarget, dist, params, NatafIntMethod = "GH", NoEval = 9,
  polydeg = 6, ...)
}
\arguments{
\item{s2srtarget}{A k-dimensional vector with the lag-1 season-to-season correlation coefficients.}

\item{dist}{A k-dimensional string vector indicating the quantile function of the target marginal distributions (i.e., the ICDFs).}

\item{params}{A k-dimensional named list with the parameters of the target distributions.}

\item{NatafIntMethod}{A string ("GH", "Int", or "MC") indicating the integration method to resolve the Nataf integral.}

\item{NoEval}{A scalar indicating (default: 9) the number of evaluation points for the integration methods.}

\item{polydeg}{A scalar indicating the order of the fitted polynomial. If polydeg=0, then another curve is fitted.}

\item{...}{Additional named arguments for the selected "NatafIntMethod" method.}
}
\value{
A list with the parameters of the cyclostationary auxiliary Gaussian PAR(1) model.
}
\description{
Estimation of parameters of the auxiliary PAR(1) model to simulate the auxiliary cyclostationary Gaussian process to establish the target season-to-season correlation structure.
}
\note{
Avoid the use of the "GH" method (i.e., NatafIntMethod='GH'), when the marginal(s) are discrete.
}
\examples{
## Parameter estimation of a cyclostationary process with 12 seasons.
\dontrun{
rtarget<-c(0.5, 0.7, 0.6, 0.4, 0.5, 0.7, 0.8, 0.7, 0.6, 0.4, 0.5, 0.7)
NumOfSeasons=length(rtarget)

FXs<-rep('qmixed',NumOfSeasons)
PFXs<-vector("list",NumOfSeasons)
PFXs<-lapply(PFXs,function(x) x<-list(p0=0.4,Distr=qexp,rate=0.5))

SPARTApar<-EstSPARTA(s2srtarget = rtarget, dist = FXs,
params = PFXs, NatafIntMethod = 'GH', NoEval = 9, polydeg = 8)
}
}
