% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatafGH.R
\name{NatafGH}
\alias{NatafGH}
\title{Solve the Nataf integral with the Gauss-Hermite integration method}
\usage{
NatafGH(rho, fx, fy, paramlistfx, paramlistfy, nodes = 21, prune = 0)
}
\arguments{
\item{rho}{A scalar or vector of correlation coefficients (i.e., in seq(from=0, to=1, by=0.1)).}

\item{fx}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{fy}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{paramlistfx}{A named list with the parameters of the distribution.}

\item{paramlistfy}{A named list with parameters of the distribution.}

\item{nodes}{A scalar indicating the number of nodes for Gauss-Hermite integration (default: nodes=21).}

\item{prune}{A scalar in (0,1) indicating the percentage of pruning for Gauss-Hermite integration (default: nodes=0).}
}
\value{
A vector of correlation coefficients in the actual domain
}
\description{
Estimation of the resulting (i.e., in the actual domain) correlation coefficients,
given the equivalent correlation coefficients (i.e., in the Gaussian domain).
}
\note{
Avoid the use of this function, when the marginal(s) are discrete.
based on the blog-post of Matthew S. Shotwell (https://biostatmatt.com/archives/2754).
}
\examples{
## The case of two identrical Gamma distributions, with shape=1 and scale=1.
\dontrun{
fx=fy='qgamma'
pfx=pfy=list(shape=1, scale=1)
rhoz=seq(from=0, to=1 , by=0.05)
rhox=NatafGH(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx,
paramlistfy = pfy, nodes = 10, prune = 0)
plot(rhoz,rhox); abline(0,1)

## The case of two identrical zero-inflated (i.e., mixed) distributions,
with p0=0.7 a Gamma distribution
## for the continuous part with shape=1 and scale=1.

fx=fy='qmixed'
pfx=pfy=list(Distr=qgamma, p0=0.7, shape=0.5, scale=1)
rhoz=seq(from=0, to=1 , by=0.05)
rhox=NatafGH(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx,
paramlistfy = pfy, nodes = 21, prune = 0)
plot(rhoz,rhox); abline(0,1)
}
}
