% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatafInt.R
\name{NatafInt}
\alias{NatafInt}
\title{Solve the Nataf integral with an alternative integration method}
\usage{
NatafInt(rho, fx, fy, paramlistfx, paramlistfy)
}
\arguments{
\item{rho}{A scalar or vector of correlation coefficients (i.e., in seq(from=0, to=1, by=0.1)).}

\item{fx}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{fy}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{paramlistfx}{A named list with the parameters of the distribution.}

\item{paramlistfy}{A named list with parameters of the distribution.}
}
\value{
A vector of correlation coefficients in the actual domain.
}
\description{
Estimation of the resulting (i.e., in the actual domain) correlation coefficients,
given the equivalent correlation coefficients (i.e., in the Gaussian domain).
}
\examples{
## The case of two identrical Gamma distributions, with shape=1 and scale=1.
\dontrun{
fx=fy='qgamma'
pfx=pfy=list(shape=1, scale=1)
rhoz=seq(from=0, to=1 , by=0.2)
rhox=NatafInt(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
plot(rhoz,rhox); abline(0,1)

## The case identrical Bernoulli distributions, with size=1 and prob=0.3.
fx=fy='qbinom'
pfx=pfy=list(size=1, prob=0.3)
rhoz=seq(from=0, to=1 , by=0.2)
rhox=NatafInt(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
plot(rhoz,rhox); abline(0,1)

## The case of two identrical zero-inflated (i.e., mixed) distributions,
## with p0=0.7 a Gamma distribution
## for the continuous part with shape=1 and scale=1.

fx=fy='qzi'
pfx=pfy=list(Distr=qgamma, p0=0.7, shape=1, scale=1)
rhoz=seq(from=0, to=1 , by=0.2)
rhox=NatafInt(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
plot(rhoz,rhox);abline(0,1)
# compare with the Gauss-Hermite integration method
rhox=NatafGH(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy, nodes = 21)
points(rhoz,rhox,col='red',pch=19)
}
}
