% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NatafMC.R
\name{NatafMC}
\alias{NatafMC}
\title{Solve the Nataf integral with the Monte-Carlo method}
\usage{
NatafMC(rho, fx, fy, paramlistfx, paramlistfy, MCsize = 10^5)
}
\arguments{
\item{rho}{A scalar or vector of correlation coefficients (i.e., in seq(from=0, to=1, by=0.1)).}

\item{fx}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{fy}{A string indicating the quantile function of the distribution (i.e., the ICDF).}

\item{paramlistfx}{A named list with the parameters of the distribution.}

\item{paramlistfy}{A named list with parameters of the distribution.}

\item{MCsize}{A scalar determining the number of Monte-Carlo trials (default: 10^-5)}
}
\value{
A vector of correlation coefficients in the actual domain.
}
\description{
Estimation of the resulting (i.e., in the actual domain) correlation coefficients,
given the equivalent correlation coefficients (i.e., in the Gaussian domain).
}
\examples{
## The case of two identical Gamma distributions, with shape=1 and scale=1.
\dontrun{
fx=fy='qgamma'
pfx=pfy=list(shape=1, scale=1)
rhoz=seq(from=0, to=1 , by=0.2)
rhoxMC=NatafMC(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
plot(rhoz,rhoxMC, col='blue', pch=19); abline(0,1)
rhoxGH=NatafGH(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
points(rhoz, rhoxGH, col='red', pch=5)

## The case identrical Bernoulli distributions, with size=1 and prob=0.3.

fx=fy='qbinom'
pfx=pfy=list(size=1, prob=0.3)
rhoz=seq(from=0, to=1 , by=0.2)
rhoxMC=NatafMC(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
plot(rhoz,rhoxMC, col='blue', pch=19); abline(0,1)
}

\dontrun{
rhoxInt=NatafInt(rho = rhoz, fx = fx, fy = fy, paramlistfx = pfx, paramlistfy = pfy)
points(rhoz, rhoxInt, col='red', pch=5)
}
}
