% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimARTAp.R
\name{SimARTAp}
\alias{SimARTAp}
\title{Simulation of the target stationary process using the ARTA(p) model.}
\usage{
SimARTAp(ARTApar, burn = 1000, steps = 10000, stand = F)
}
\arguments{
\item{ARTApar}{A list containing the parameters of the model. The list is constructed by the function "EstARTAp".}

\item{burn}{A scalar specifying the length of burn-out sample.}

\item{steps}{A scalar specifying the length of the time series to be generated.}

\item{stand}{A boolean (T or F) indicating whether to standardize (or not) the auxiliary Gaussian time series prior to their mapping to the actual domain. The default value is FALSE.}
}
\value{
A list of the 3 generated time series (in vector format):
X: The final time series at the actual domain with the target marginal distribution and correlation structure;
Z: The auxiliary Gaussian time series at the Gaussian domain and;
U: The auxiliary uniform time series at the Copula domain (i.e., in [0,1]).
}
\description{
Simulation of the target stationary process using an ARTA(p) model to simulate the auxiliary Gaussian process and establish the target correlation structure.
}
\examples{

## Simulation of a process with a zero-inflated (i.e., mixed)
## marginal distribution, with p0=0.9, and a Gamma distribution
## for the continuous part with shape=0.1 and scale=1.
## In this case, the target autocorrelation strucure is from
## the CAS ACS with b=1 and k=0.6.
\dontrun{
ACF=acsCAS(param=c(1, 0.6), lag=100, var=1)
fx='qmixed'
pfx=list(Distr=qgamma, p0=0.9, shape=0.1, scale=1)

ARTApar=EstARTAp(ACF=ACF, maxlag=0, dist=fx, params=pfx,
NatafIntMethod ='GH', NoEval=9, polydeg=8)

Sim=SimARTAp(ARTApar = ARTApar, burn = 1000, steps = 10^5, stand = 0)
acf(Sim$X)
lines(0:(length(ACF)-1), ACF)
plot(Sim$X[1:1000], type='l', col='red')

## Simulation of a process with a bernoulli marginal distribution,
## with size=1, and prob=0.2.
## In this case, the target autocorrelation strucure
## is from an fGn process (i.e., Hurst) with H=0.7.

ACF=acsHurst(H=0.7, lag=500, var=1)
fx='qbinom'
pfx=list(size=1, prob=0.2)

ARTApar=EstARTAp(ACF=ACF, maxlag=0, dist=fx, params=pfx,
NatafIntMethod ='Int', NoEval=9, polydeg=0)

Sim=SimARTAp(ARTApar = ARTApar, burn = 1000, steps = 10^5, stand = 0)
acf(Sim$X)
lines(0:(length(ACF)-1), ACF)
plot(Sim$X[1:1000], type='l', col='red')

## Simulation of a process with a Beta marginal distribution,
## with shape=2, and shape2=10.
## In this case, the target autocorrelation strucure is from
## an fGn process (i.e., Hurst) with H=0.7.

ACF=acsHurst(H=0.7, lag=500, var=1)
fx='qbeta'
pfx=list(shape1=2, shape2=10)

ARTApar=EstARTAp(ACF=ACF, maxlag=0, dist=fx, params=pfx,
NatafIntMethod ='Int', NoEval=9, polydeg=0)

Sim=SimARTAp(ARTApar = ARTApar, burn = 1000, steps = 10^5, stand = 0)
acf(Sim$X)
lines(0:(length(ACF)-1), ACF)
plot(Sim$X[1:1000], type='l', col='red')
}
}
