% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimSPARTA.R
\name{SimSPARTA}
\alias{SimSPARTA}
\title{Simulation of the target cyclostationary process using the SPARTA model of order 1.}
\usage{
SimSPARTA(SPARTApar, steps = 1000, stand = 0)
}
\arguments{
\item{SPARTApar}{A list containing the parameters of the model. The list is constructed by the function "EstSPARTA".}

\item{steps}{A scalar specifying the length of the time series to be generated.}

\item{stand}{A boolean (T or F) indicating whether to standardize (or not) the auxiliary Gaussian time series prior to their mapping to the actual domain. The default value is FALSE.}
}
\value{
A list of 3 generated time series (in matrix format - i.e.,  matrix of dimensions k x m, where m denotes the number of sub-seasons and k the number of periods.):
X: The final time series at the actual domain with the target marginal distribution and correlation structure;
Z: The auxiliary Gaussian time series at the Gaussian domain and,
U: The auxiliary uniform time series at the Copula domain (i.e., in [0,1]).
}
\description{
Simulation of the target cyclostationary process using a PAR(p) model (i.e., the SPARTA model of order 1) to simulate the auxiliary cyclostationary Gaussian process to establish the target season-to-season correlation structure.
}
\examples{
## Simulation of cyclostationary process with 12 seasons and zero-inflated marginal distributions.
\dontrun{
rtarget<-c(0.5, 0.7, 0.6, 0.4, 0.5, 0.7, 0.8, 0.7, 0.6, 0.4, 0.5, 0.7)
NumOfSeasons=length(rtarget)
FXs<-rep('qmixed',NumOfSeasons)
PFXs<-vector("list",NumOfSeasons)
PFXs<-lapply(PFXs,function(x) x<-list(p0=0.4, Distr=qexp, rate=0.5))

SPARTApar<-EstSPARTA(s2srtarget = rtarget, dist = FXs, params = PFXs,
NatafIntMethod = 'GH', NoEval = 9, polydeg = 8)

sim<-SimSPARTA(SPARTApar = SPARTApar, steps=100000, stand=0)
s2scor(sim$X)
plot(s2scor(sim$X)); lines(rtarget, col='red')
}
}
